\name{landscape_change}
\alias{landscape_change}
\docType{data}
\title{
Landscape loosing 5\% of patches per time step
}
\description{
This dataset is a list of 100 landscapes with a loss of 5\% of each patch's area at each time step. The first landscape is the sample empty landscape.
}
\format{
List of 100 data frames, that represent the evolution of the landscape during 100 time steps.
}
\examples{
data(landscape_change)
}
\keyword{datasets}
