\name{MetaLandSim-internal}
\alias{merge_order}
\title{
Internal functions for the MetaLandSim package.
}
\description{
Internal functions for the MetaLandSim package 
}
\details{
These are not to be called by the user.}
\source{
Coded by Tal Galili. URL: http://www.r-statistics.com/2012/01/merging-two-data-frame-objects-while-preserving-the-rows-order/
}

