\name{MetaLandSim-package}
\alias{MetaLandSim-package}
\alias{MetaLandSim}
\docType{package}
\title{
Landscape, metapopulation and range expansion simulation
}
\description{
The package MetaLandSim allows the generation of random landscapes, represented as graphs, the simulation of landscape dynamics, metapopulation dynamics and range expansion into empty landscapes.
The package was developed as part of the Ph.D. thesis of Frederico Mestre (SFRH/BD/73768/2010) and included in the project NETPERSIST (PTDC/AAG-MAA/3227/2012) both financed by the Fundacao para a Ciencia e Tecnologia.
}
\details{
\tabular{ll}{
Package: \tab MetaLandSim\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-02-02\cr
License: GPL (>=2)\cr
}
}
\author{
Frederico Mestre, Fernando Canovas, Ricardo Pita, Antonio Mira and Pedro Beja.

Maintainer: Frederico Mestre <mestre.frederico@gmail.com>

}
\references{
Mestre F, Canovas F, Pita R, Mira A and Beja P (submitted) 'MetaLandSim: tools for simulating metapopulation persistence in dynamic landscapes and range expansion'
}
