\name{MetaLandSim.GUI}
\alias{MetaLandSim.GUI}
\title{
Graphic User Interface
}
\description{
User-friendly graphic user interface that allows running the main functions of the package. 
}
\usage{
MetaLandSim.GUI()
}
\value{
Displays the graphic user interface.
}
\author{
Frederico Mestre and Fernando Canovas
}
\seealso{
\code{\link{rland.graph}}, \code{\link{species.graph}}, \code{\link{simulate_graph}}, 
\code{\link{iterate.graph}}, \code{\link{range_expansion}},\cr \code{\link{range_raster}} 
}
\examples{

\dontrun{
#In order to display the GUI:

MetaLandSim.GUI()
}
}
