\name{rland}
\alias{rland}
\docType{data}
\title{
Random landscape
}
\description{
Sample random landscape graph, object of class 'landscape'.  It has 60 patches and the landscape mosaic has 1000 meters of side.
}
\usage{data(rland)}
\format{A list with the following elements:
\itemize{
 \item mapsize - landscape mosaic side length, in meters.
 \item minimum.distance - minimum distance between patches centroids).
 \item mean.area - mean area, in hectares.
 \item SD.area - standard deviation of the area.
 \item number.patches - number of patches.
 \item dispersal - mean dispersal ability of the species.
 \item nodes.characteristics - data frame with the characteristics of each patch.
}
}
\examples{

data(rland)

}
\keyword{datasets}
