\name{sim.area}
\alias{sim.area}
\docType{data}
\title{
Vector of the areas for each site; here, 100 sites
}
\description{
By loading simulatedifm this object is loaded.
}
\format{
  \describe{
    \item{\code{sim.area}}{Vector of the areas for each site; here, 100 sites.}
  }
}
\details{
This dataset was created using the code included in \bold{Examples}, in \code{simulatedifm}.
}
\keyword{datasets}
\keyword{ifm}
