\name{sim.det.20}
\alias{sim.det.20}
\docType{data}
\title{
Array corresponding to nsites x nyears x nvisits
}
\description{
By loading simulatedifm this object is loaded.
}
\format{
  \describe{
    \item{\code{sim.det.20}}{100 x 10 x 3 array corresponding to nsites x nyears x nvisits. Data simulated with year-specific detection probabilities equal to 0.4,0.6,0.2,0.9,0.3,0.4,0.6,0.2,0.9,0.3.}
  }
}
\details{
This dataset was created using the code included in \bold{Examples}, in \code{simulatedifm}.
}
\keyword{datasets}
\keyword{ifm}