\name{sim.distance}
\alias{sim.distance}
\docType{data}
\title{
Distance matrix between sampling sites (nsite x nsite).
}
\description{
By loading simulatedifm this object is loaded.
}
\format{
  \describe{
    \item{\code{sim.distance}}{nsite x nsite distance matrix.}
  }
}
\details{
This dataset was created using the code included in \bold{Examples}, in \code{simulatedifm}.
}
\keyword{datasets}
\keyword{ifm}
