\name{z.sim.20.fa}
\alias{z.sim.20.fa}
\docType{data}
\title{
Occupancy data containing false absences
}
\description{
By loading simulatedifm this object is loaded.
}
\format{
  \describe{
    \item{\code{z.sim.20.fa}}{nyear x nsite occupancy data containing false absences, which can be used to explore the bias of ifm.missing.MCMC and ifm.naive.MCMC when there is imperfect detection.}
  }
}
\details{
This dataset was created using the code included in \bold{Examples}, in \code{simulatedifm}.
}
\keyword{datasets}
\keyword{ifm}