\name{z.sim.20}
\alias{z.sim.20}
\docType{data}
\title{
Occupancy data generated with perfect detection with approximately 20\% of data missing at random.
}
\description{
By loading simulatedifm this object is loaded.
}
\format{
  \describe{
    \item{\code{z.sim.20}}{nyear x nsite occupancy data generated with perfect detection with approximately 20\% of data missing at random.}
  }
}
\details{
This dataset was created using the code included in \bold{Examples}, in \code{simulatedifm}.
}
\keyword{datasets}
\keyword{ifm}
