\name{z.sim}
\alias{z.sim}
\docType{data}
\title{
Occupancy data generated with perfect detection.
}
\description{
By loading simulatedifm this object is loaded.
}
\format{
  \describe{
    \item{\code{z.sim}}{nyear x nsite occupancy data generated with perfect detection.}
  }
}
\details{
This dataset was created using the code included in \bold{Examples}, in \code{simulatedifm}.
}
\keyword{datasets}
\keyword{ifm}