\name{plot_expansion}
\alias{plot_expansion}
\title{
Graphical display of the expansion
}
\description{
Plots the expansion object.
}
\usage{
plot_expansion(exp)
}
\arguments{
  \item{exp}{
Object of class 'expansion'.
}
}
\value{
Graphical display of the 'expansion' class.
}
\author{
Frederico Mestre and Fernando Canovas
}
\seealso{
\code{\link{expansion}}
}
\examples{

data(rg_exp)

plot_expansion(exp=rg_exp)

}
